+++ nekoaisle-cursor/src/extension.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import SelectWord from './SelectWord';
 import GotoLine from './GotoLine';
 import MarkJump from './MarkJump';
+import Cursor from './Cursor';
 
 /**
  * エクステンション活性化
@@ -12,6 +13,7 @@ export function activate(context: vscode.ExtensionContext) {
 	new SelectWord(context);
   new GotoLine(context);
   new MarkJump(context);
+  new Cursor(context);
 }
 
 /**